+++ test/fork_test.c
@@ -29,7 +29,7 @@
 static pid_t child_pid = -1;
 
 static void basic_child_func(void* arg) {
-  KEXPECT_EQ(0x1234, (uint32_t)arg);
+  KEXPECT_EQ(0x1234, (intptr_t)arg);
   char cwd[VFS_MAX_PATH_LENGTH];
   KLOG("child proc:  id: %d  arg: %p\n", proc_current()->id, arg);
   vfs_getcwd(cwd, VFS_MAX_PATH_LENGTH);
@@ -139,7 +139,7 @@ static void reparent_zombie_to_root_test(void) {
 static void multi_child_func(void* arg) {
   scheduler_yield();
   scheduler_yield();
-  proc_exit((int)arg);
+  proc_exit((intptr_t)arg);
 }
 
 static void multi_child_test(void) {
